+++ vad-filter-onnx/vad/vad-model.cc
@@ -81,7 +81,7 @@ void VadModel::reset() {
 
 void VadModel::on_voice_start() {
     // Precise start: current - consecutive speech frames - padding
-    int lookback_speech_frames = window_detector_->num_right_ones();
+    int lookback_speech_frames = static_cast<int>(window_detector_->num_right_ones());
     int lookback_speech_samples = lookback_speech_frames * frame_shift_;
     start_ = current_ - lookback_speech_samples - left_padding_samples_;
     start_ = std::max(last_end_, start_);
@@ -96,7 +96,7 @@ void VadModel::on_voice_start() {
 
 void VadModel::on_voice_end() {
     // Precise end: current - consecutive silence frames + padding
-    int lookback_silence_frames = window_detector_->num_right_zeros();
+    int lookback_silence_frames = static_cast<int>(window_detector_->num_right_zeros());
     int lookback_silence_samples = lookback_silence_frames * frame_shift_;
     end_ = current_ - lookback_silence_samples + right_padding_samples_;
     end_ = std::min(end_, current_);
